feat: add measure2 and measure3 element methods
The measure computation is done in the measure module.

For now I chose to use rust arrays, but it might be better to use
nalgebra types.

As nalgebra only offers owned data struct, element coords (the ones in
once_cell) could be in nalgebra format, but the type should not leak to
the user.

All this need benchmarking.